# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/cart": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Просмотр товаров в корзине",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CartDTO"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Добавление товара в корзину",
                "parameters": [
                    {
                        "description": "Добавить товар в козрину",
                        "name": "AddProductCartDTO",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddProductCartDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CartDTO"
                        }
                    }
                }
            }
        },
        "/api/v1/cart/product/{id}": {
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "carts"
                ],
                "summary": "Удаление товара из корзины",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ид товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CartDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AddProductCartDTO": {
            "type": "object",
            "required": [
                "product_id",
                "quantity"
            ],
            "properties": {
                "product_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dto.CartDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "in_cart": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.InCart"
                    }
                }
            }
        },
        "dto.InCart": {
            "type": "object",
            "properties": {
                "product": {
                    "$ref": "#/definitions/dto.ProductDTO"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
...
        "dto.ProductDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
    }
}